csv: integrate a simple csv parser / reader
This is a one-file class is said to reliably parse and write
a CSV file. It has no dependency on external libraries and does
the job well.

We introduce a CustomerData class that holds the customer data that
is parsed from the CSV file. The class has members in the most relevant
types to hold the various field values.

The demo main method reads a CSV file, constructs CustomerData instances
from it and further gets the path to which the customer-data needs to be
written back in CSV format.